requirements.txt
# Core
aiohttp>=3.9.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
readability-lxml>=0.8.1
pydantic>=2.5.0

# CLI
typer[all]>=0.9.0
rich>=13.0.0

# Data
pandas>=2.1.0
pyarrow>=14.0.0

# Images
Pillow>=10.0.0
imagehash>=4.3.0

# Logging
structlog>=23.0.0

# Encoding
chardet>=5.0.0

# Content dedup
simhash>=2.1.0

# Dynamic rendering (optional)
playwright>=1.40.0

# API (optional)
fastapi>=0.104.0
uvicorn>=0.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.23.0
aioresponses>=0.7.6
